TEMPLATE: add ownership and fix typos
A literature review should have a single owner. That's usually the person who originally did the literature review. They should be willing to own it. They should also answer any issues on it with the same SLA expected of other elm-community project owners. And I do mean it should be a *single* owner. If it's a team effort, the team should specify someone to take point for communication. They can list the rest of the team elsewhere in the document for credit's sake, or keep a human-meaningful changelog.

But GitHub issues and PRs are not always the best place for feedback. In that case, the owner should specify their preferred method of communication. Probably something on the Elm Slack, maybe elm-discuss or Reddit. Quick questions and real-time discussion that results in the proposal moving forward should happen in that space.

Also, I fixed some typos. Markdown interprets alphanumeric values inside lt/gt signs as HTML tags so the project title was invisible.